refactor(application): squashed
refactor(gui/backend): rename subdir
- Core as a name is less confusing
- fyne-gui allows to support different UIs in the future
refactor(fyne-gui): rename package
refactor(core): rename package
refactor(core): overhaul
- renaming packages, events
- restricted struct field and method access where its not necessary
- remove unneccessary methods
refactor(main): imported package names
refactor(node): event bind callback signature
refactor(modal): wrap popup in struct
refactor(fyne-gui): overhaul
- package renaming
- file structuring
- adapt eventbus changes
- remove mutexes
style(comment): "backend" to "core" renaming
