+++ fyne-gui/Console.go
@@ -1,31 +1,28 @@
-package gui
+package fynegui
 
 import (
 	"distributed-sys-emulator/bus"
-	"sync"
 
 	"fyne.io/fyne/v2"
 	"fyne.io/fyne/v2/container"
 	"fyne.io/fyne/v2/widget"
 )
 
+// Declare conformance with the Component interface
+var _ Component = (*Console)(nil)
+
 type Console struct {
 	*fyne.Container
 }
 
-// Declare conformance with the Component interface
-var _ Component = (*Console)(nil)
-
-func NewConsole(eb *bus.EventBus) *Console {
+func NewConsole(eb bus.EventBus) *Console {
 	c := container.NewBorder(nil, nil, nil, nil)
-	cmu := sync.Mutex{}
 
 	var nodeCnt int
 	var outputs []string
 
 	// depending on the outputs, create console boxes
 	refresh := func() {
-		cmu.Lock()
 		c.RemoveAll()
 		grid := container.NewGridWithColumns(nodeCnt)
 		for i := 0; i < nodeCnt; i++ {
@@ -34,12 +31,11 @@ func NewConsole(eb *bus.EventBus) *Console {
 		}
 		c.Add(grid)
 		c.Refresh()
-		cmu.Unlock()
 	}
 
 	// update node count and output slice size as required
-	eb.Bind(bus.NetworkNodeCntChangeEvt, func(e bus.Event) {
-		nodeCnt = e.Data.(int)
+	eb.Bind(bus.NetworkResizeEvt, func(resizeData bus.NetworkResize) {
+		nodeCnt = resizeData.Cnt
 		if nodeCnt > len(outputs) {
 			for i := 0; i <= nodeCnt-len(outputs); i++ {
 				outputs = append(outputs, "")
@@ -49,8 +45,7 @@ func NewConsole(eb *bus.EventBus) *Console {
 	})
 
 	// update outputs
-	eb.Bind(bus.OutputChanged, func(e bus.Event) {
-		out := e.Data.(bus.Output)
+	eb.Bind(bus.NodeOutputLogEvt, func(out bus.NodeLog) {
 		outputs[out.NodeId] = out.Str
 		refresh()
 	})
